Readme

+++ ejercicio5.js
@@ -1,5 +1,5 @@
-const peso = parseInt (prompt("¿cual es tu peso?"));
-const altura = parseInt (prompt("¿cual es tu altura?"));
+const peso = parseFloat (prompt("¿cual es tu peso?"));
+const altura = parseFloat (prompt("¿cual es tu altura?"));
 
 const imc = peso/altura**2;
